+++ src/main/kotlin/moe/kurenai/tdlight/request/message/EditMessageRequest.kt
@@ -3,7 +3,7 @@ package moe.kurenai.tdlight.request.message
 interface EditMessageRequest {
 
     val chatId: String?
-    val messageId: Int?
+    val messageId: Long?
     val inlineMessageId: String?
 
 }
