+++ src/main/kotlin/moe/kurenai/tdlight/request/message/Reply.kt
@@ -1,6 +1,6 @@
 package moe.kurenai.tdlight.request.message
 
 interface Reply {
-    var replyToMessageId: Int?
+    var replyToMessageId: Long?
     var allowSendingWithoutReply: Boolean?
 }
