+++ src/main/kotlin/moe/kurenai/tdlight/request/message/SendDocument.kt
@@ -21,7 +21,7 @@ data class SendDocument(
     var thumb: InputFile? = null
     var disableContentTypeDetection: Boolean? = null
     override var parseMode: String? = null
-    override var replyToMessageId: Int? = null
+    override var replyToMessageId: Long? = null
     override var allowSendingWithoutReply: Boolean? = null
     override var disableNotification: Boolean? = null
     override var sendAt: Long? = null
