+++ src/main/kotlin/moe/kurenai/tdlight/request/message/SendSticker.kt
@@ -16,7 +16,7 @@ data class SendSticker(
     val sticker: InputFile,
 ) : Request<ResponseWrapper<Message>>(), SendMessageRequest, MediaRequest, Reply, WithCaption, WithReplyMarkup {
     override var parseMode: String? = null
-    override var replyToMessageId: Int? = null
+    override var replyToMessageId: Long? = null
     override var allowSendingWithoutReply: Boolean? = null
     override var disableNotification: Boolean? = null
     override var sendAt: Long? = null
